Use Memcache while indexing Evergreen
The in-memory cache doesn't share data between threads, making it
of little use for the multi-threaded eg authority control fetcher.

In practice, this meant that I ended up spamming the Evergreen server
for authority records that had already been fetched and cached in a
different thread -- millions of useless calls!

Let's use memcache instead.  The memcache connection is only necessary
for indexing Evergreen.